requirements.txt
cuda-python==11.7.0
einops==0.4.1
glob2==0.7
idna==3.3
kiwisolver==1.4.2
lmdb==1.3.0
msgpack==1.0.4
nvidia-pyindex==1.0.9
onnx==1.10.2
onnx-simplifier==0.3.6
onnxoptimizer==0.2.6
onnxruntime==1.10.0
Pillow==9.1.0
polygraphy==0.36.2
protobuf==3.19.1
pyarrow==8.0.0
pycuda==2021.1
six==1.15.0
torch==1.11.0
torchvision==0.12.0
tqdm==4.64.0